Move logic for querying openmetrics into an openmetrics source.
