+++ sources/openmetrics/prometheus_test.go
@@ -1,0 +1,26 @@
+package openmetrics
+
+import (
+	"regexp"
+	"testing"
+
+	dto "github.com/prometheus/client_model/go"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestShouldExportMetric(t *testing.T) {
+	metric1Name := "metric1Name"
+
+	mf := dto.MetricFamily{
+		Name: &metric1Name,
+	}
+
+	ignoredMetrics1 := []*regexp.Regexp{
+		regexp.MustCompile(".*0.*"),
+		regexp.MustCompile(".*1.*"),
+	}
+	ignoredMetrics2 := []*regexp.Regexp{regexp.MustCompile(".*2.*")}
+
+	assert.False(t, shouldExportMetric(mf, ignoredMetrics1))
+	assert.True(t, shouldExportMetric(mf, ignoredMetrics2))
+}
